`JSONRPCParam` enum implementation.
Enum to compose request to the node params.

In most cases request params are passed to Ethereum JSON RPC request as array of mixed type values, such as `[12,"this",true]`,
thus this is not appropriate API design we have what we have.

Meanwhile swift don't provide strict way to compose such array it gives some hacks to solve this task
and one of them is using `RawRepresentable` protocol.

Conforming this protocol gives designated type ability to represent itself in `String` representation in any way.

So in our case we're using such to implement custom `encode` method to any used in node request params types.