To create `JSONRPCrequest` parameters should conforms `JSONRPCParameter` protocol.
Also there's added `RPCParameter` enum which implements internal encoding logic of value of any conformed `JSONRPCParameter` type.

Unfortunately there's no way in swift yet to encode generic types. Therefore there's no way to made both `JSONRPCParameter` & `RPCParameter` types extendable.

So yet there's a manual restriction for supported types, **must** conforms rather `JSONRPCParameter` protocol for base types, and `JSONParameterElement` protocol for types that are stored in arrays. The latter protocol is internal to deny conforming to it by a library user.

`JSONRPCParameter` itself could be one to conform with by a user custom type, since there's a lack of making this type internal due to `RawRepresentable` protocol restrictions. So there's nothing we can do, but a leaving a documentation with denying to do this.